+++ tests/cache/utils.js
@@ -16,13 +16,11 @@ const tu = require('../testUtils');
 const testStartTime = new Date();
 
 module.exports = {
-  forceDelete: (done) => {
-    tu.forceDelete(tu.db.Sample, testStartTime)
+  forceDelete: (done) => tu.forceDelete(tu.db.Sample, testStartTime)
     .then(() => tu.forceDelete(tu.db.Aspect, testStartTime))
     .then(() => tu.forceDelete(tu.db.Subject, testStartTime))
     .then(() => tu.forceDelete(tu.db.Profile, testStartTime))
     .then(() => tu.forceDelete(tu.db.User, testStartTime))
     .then(() => done())
-    .catch(done);
-  },
+    .catch(done),
 };
